Algorithm: Implemented ‘allocation strategy’
Instead of assigning SOLL just 1 place and expanding afterwards , it now determines a strategy beforehand: optimistic or conservative.

The latter is the old situation, the former gives everyone who’s interested a maximum of 2 places right off the bat. This also applies to VPH with 1 place trying to expand.

People with `minimum > 2` can still be rejected during the expansion phase.

Solves #370